chore: strict script error handling

+++ .setup.d/darwin/00-variables.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+set -Eeuo pipefail
+
+: "${ARCH:?unknown architecture}"
+: "${SYSTEM:?unknown system}"
+
 sudo launchctl config user path $PATH
 launchctl setenv PATH $PATH
 killall Dock
